requirements.txt
openai
google-generativeai
sounddevice
scipy
python-dotenv
Django
djangorestframework